Wrote a simple program to optimize the field multiplication coefficients. Include early results
thoughts here are: I'd like common cases (i.e. the first 50 or so fields) not to wrap in 64-bit arithmetic.
This allows me to use plain old normal math to ensure that the coefficients are relatively prime
 - of course I guess I could just use primes themselves...
Furthermore, I want coefficients to be relatively large, so that I'm using as much of my 64-bit entropy
as possible to encourage good mixing.  In practice this means I want my multiplicative factor to be not\
too far from 2^64/2^32/100 say, i.e. not too much smaller than 50 million.